api/requirements.txt
asgiref
Django==4.0
django-cors-headers
djangorestframework
djangorestframework_simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
boto3
Pillow